pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rawdata"
version = "0.1.0"
description = "Spreadsheet-as-database with Git sync and agent-friendly CLI"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "gitpython>=3.1",
    "pandas>=2.0",
    "anthropic>=0.25",
    "rich>=13.0",
]

[project.scripts]
rd = "rawdata.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["rawdata"]
